+++ packages/expo-location/ios/EXLocation/Requesters/EXForegroundPermissionRequester.m
@@ -22,8 +22,8 @@ static SEL whenInUseAuthorizationSelector;
 
 - (void)requestLocationPermissions
 {
-  if ([EXBaseLocationRequester isConfiguredForWhenInUseAuthorization] && [self.locMgr  respondsToSelector:whenInUseAuthorizationSelector]) {
-    ((void (*)(id, SEL))objc_msgSend)(self.locMgr, whenInUseAuthorizationSelector);
+  if ([EXBaseLocationRequester isConfiguredForWhenInUseAuthorization] && [self.locationManager  respondsToSelector:whenInUseAuthorizationSelector]) {
+    ((void (*)(id, SEL))objc_msgSend)(self.locationManager, whenInUseAuthorizationSelector);
   } else {
     self.reject(@"ERR_LOCATION_INFO_PLIST", @"The `NSLocationWhenInUseUsageDescription` key must be present in Info.plist to be able to use geolocation.", nil);