+++ packages/expo-location/ios/EXLocation/Requesters/EXLocationPermissionRequester.m
@@ -23,10 +23,10 @@ static SEL whenInUseAuthorizationSelector;
 
 - (void)requestLocationPermissions
 {
-  if ([EXBaseLocationRequester isConfiguredForAlwaysAuthorization] && [self.locMgr respondsToSelector:alwaysAuthorizationSelector]) {
-    ((void (*)(id, SEL))objc_msgSend)(self.locMgr, alwaysAuthorizationSelector);
-  } else if ([EXBaseLocationRequester isConfiguredForWhenInUseAuthorization] && [self.locMgr respondsToSelector:whenInUseAuthorizationSelector]) {
-    ((void (*)(id, SEL))objc_msgSend)(self.locMgr, whenInUseAuthorizationSelector);
+  if ([EXBaseLocationRequester isConfiguredForAlwaysAuthorization] && [self.locationManager respondsToSelector:alwaysAuthorizationSelector]) {
+    ((void (*)(id, SEL))objc_msgSend)(self.locationManager, alwaysAuthorizationSelector);
+  } else if ([EXBaseLocationRequester isConfiguredForWhenInUseAuthorization] && [self.locationManager respondsToSelector:whenInUseAuthorizationSelector]) {
+    ((void (*)(id, SEL))objc_msgSend)(self.locationManager, whenInUseAuthorizationSelector);
   } else {
     self.reject(@"E_LOCATION_INFO_PLIST", @"One of the `NSLocation*UsageDescription` keys must be present in Info.plist to be able to use geolocation.", nil);