+++ packages/expo-location/src/Location.ts
@@ -155,6 +155,7 @@ export async function reverseGeocodeAsync(
 
 /**
  * Gets the current state of location permissions.
+ * @deprecated Use `getForegroundPermissions` or `getBackgroundPermissions` instead.
  */
 export async function getPermissionsAsync(): Promise<LocationPermissionResponse> {
   console.warn(
@@ -165,6 +166,7 @@ export async function getPermissionsAsync(): Promise<LocationPermissionResponse>
 
 /**
  * Requests the user to grant location permissions.
+ * @deprecated Use `requestForegroundPermissionsAsync` or `requestBackgroundPermissionsAsync` instead.
  */
 export async function requestPermissionsAsync(): Promise<LocationPermissionResponse> {
   console.warn(